Backend/backend/requirements.txt
appwrite
fastapi
uvicorn[standard]
python-dotenv
pydantic
email-validator
python-jose[cryptography]
httpx
